+++ src/main/java/me/ajlane/geo/repast/succession/pathway/EnvrConsequent.java
@@ -1,5 +1,7 @@
 package me.ajlane.geo.repast.succession.pathway;
 
+import java.io.Serializable;
+
 /**
  * Stores the information about the target state resulting from a particular combination of
  * environmental conditions stored in an EnvironmentalAntecedent
@@ -7,8 +9,8 @@ package me.ajlane.geo.repast.succession.pathway;
  * @param <T> type of the start state specifier (e.g Integer or String)
  * @author Andrew Lane
  */
-public class EnvrConsequent<T1, T2> {
-
+public class EnvrConsequent<T1, T2> implements Serializable {
+  private static final long serialVersionUID = 5284769189330757617L;
   private T1 targetState; // the land cover state this consequent represents
   private T2 transitionTime; // the number of years taken to transition to targetState